+++ test/04-html_element.js
@@ -1,0 +1,84 @@
+var assert   = require('assert'),
+    Hawkejs  = require('../index.js'),
+    hawkejs,
+    test_id = 0;
+
+describe('HTMLElement', function() {
+
+	before(function() {
+		hawkejs = new Hawkejs();
+	});
+
+	describe('#className', function() {
+		it('gets the current class attribute', function() {
+			var el = Hawkejs.Hawkejs.createElement('div');
+			el.classList.add('a');
+			el.classList.add('b');
+
+			assert.strictEqual(el.outerHTML, '<div class="a b"></div>');
+		});
+	});
+
+	describe('#id', function() {
+		it('get/set the id of the element', function() {
+
+			var el = Hawkejs.Hawkejs.createElement('div');
+
+			el.id = 'test';
+
+			assert.strictEqual(el.outerHTML, '<div id="test"></div>');
+		});
+	});
+
+	describe('#name', function() {
+		it('get/set the name of the element', function() {
+
+			var el = Hawkejs.Hawkejs.createElement('div');
+
+			el.name = 'test';
+
+			assert.strictEqual(el.outerHTML, '<div name="test"></div>');
+		});
+	});
+
+	describe('#tabIndex', function() {
+		it('get/set the tabindex of the element', function() {
+
+			var el = Hawkejs.Hawkejs.createElement('div');
+
+			el.tabIndex = 1;
+
+			assert.strictEqual(el.outerHTML, '<div tabindex="1"></div>');
+		});
+	});
+
+	describe('#value', function() {
+		it('gets/sets the value of the element, depending on the type', function() {
+
+			var input = Hawkejs.Hawkejs.createElement('input'),
+			    area = Hawkejs.Hawkejs.createElement('textarea'),
+			    div = Hawkejs.Hawkejs.createElement('div');
+
+			input.value = 'test';
+			div.value = 'test';
+
+			assert.strictEqual(input.value, 'test');
+			assert.strictEqual(div.value, 'test');
+
+			assert.strictEqual(area.value, '');
+			area.innerHTML = 'html';
+			assert.strictEqual(area.value, 'html');
+
+			area.value = 'test';
+			assert.strictEqual(area.value, 'test');
+
+			// Setting a textarea's value breaks the link with the innerText
+			assert.strictEqual(area.innerHTML, 'html');
+
+			assert.strictEqual(input.outerHTML, '<input value="test">');
+			assert.strictEqual(area.outerHTML, '<textarea>html</textarea>');
+			assert.strictEqual(div.outerHTML, '<div></div>');
+		});
+	});
+
+});